FPGA01 enablement: 24-bit CellId/NetId slot, packer shared-mux refusal, topo init rework
- Widen CellId/NetId to 24-bit slot + 8-bit generation. Previous
  16-bit slot aliased past 65k cells, masking FPGA01 cells silently.
- Packer refuses adding a child cell to a cluster when an existing
  child of the same type already occupies the same (rel_x, rel_y).
  xc7 SLICE input muxes route CE/CLK/RST through wires that resolve
  to the same routing node across z slots, so co-tile same-type
  children always conflict at the shared-mux registry.
- Rewrite initial_placement_topological as a single pass:
    1. Random shuffle for region-constrained cells.
    2. Kahn-layered barycenter for everyone else.
    3. Snap each cell to the nearest available BEL of its bucket,
       far-from-centre cells claim first.
  Removes the random pre-bind + boundary-centroid rebind pair and
  the CLB-cell-types filter; IOBs land on chip-edge BELs naturally
  via bucket membership.

+++ crates/nextpnr/src/netlist/design.rs
@@ -9,12 +9,12 @@ use super::{
 pub struct Design {
     cells: FxHashMap<IdString, CellId>,
     cell_store: Vec<Option<CellInfo>>,
-    cell_generation: Vec<u16>,
+    cell_generation: Vec<u8>,
     free_cell_slots: Vec<u32>,
 
     nets: FxHashMap<IdString, NetId>,
     net_store: Vec<Option<NetInfo>>,
-    net_generation: Vec<u16>,
+    net_generation: Vec<u8>,
     free_net_slots: Vec<u32>,
 
     pub hierarchy: FxHashMap<IdString, HierarchicalCell>,
